Added data initialisation
Added a roles.sql file that will insert roles on application startup. Also changed ddl-auto to create so that it drops/creates tables on application launch. This is useful for development but we shouldn't keep it like this. If you need to add any default table data then create an sql script for the table you want, and add inserts for all the data. Then you just need to add that file to spring.sql.init.data-locations. See  classpath:roles.sql for more info