Employee and User updated
Both classes should be related as intended now. Each user should have a employee, but employee doesn't require a user. Added test cases to figure out how that persistence is supposed to correctly function. Had to remove setters for user_id and employee_id as this leads to hibernate consistency errors. For employee this also meant removing it from the constructor, this might break things.